同步6.1release代码

+++ services/scan_service/include/scan_service_utils.h
@@ -30,7 +30,7 @@ public:
         int32_t userId, int32_t callerPid);
     static ScanErrorCode ConvertErro(const SaneStatus status);
     static std::vector<std::string> ExtractIpOrPortFromUrl(const std::string& url,
-        const char delimiter, const int32_t minTokenLength);
+        const char delimiter, const size_t minTokenLength);
 private:
     static std::unordered_map<SaneStatus, ScanErrorCode> saneStatusToScanErrorCodeMap_;
 };
